Create module for vec-based parser transformers
